frequencies loaded via initial action
