+++ src/actions/types.js
@@ -1,6 +1,5 @@
 export const UPDATE_TIP_RADII = "UPDATE_TIP_RADII";
 export const NEW_COLORS = "NEW_COLORS";
-export const INITIALISE_FREQUENCIES = "INITIALISE_FREQUENCIES";
 export const FREQUENCY_MATRIX = "FREQUENCY_MATRIX";
 export const BROWSER_DIMENSIONS = "BROWSER_DIMENSIONS";
 export const BRANCH_MOUSEENTER = "BRANCH_MOUSEENTER";
